+++ tests/fixtures/classfinder/one_class/SingleClass.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Fixture\ClassFinder;
+
+class SingleClass
+{
+}